Compile under .net9.
Everything compiles now! But does it run? Time will tell.

KSoft now has a sln-level dependency on KSoft.T4, as the latter was kept on netframework481. It cannot be a project reference, as msbuild likes to warn about it not matching KSoft's TargetFramework. So, this was the best way to get KSoft.T4 to build before KSoft. There's also some hackery in KSoft.csproj in the event the sln changes get messed up, or someone sets up a new sln for some reason.